Refactor websocket for API tests
