+++ src/test/k6/util/utils.js
@@ -11,3 +11,7 @@ export function nextAlphanumeric(length) {
 
     return randString(length, allowedChars);
 }
+
+export function nextWSSubscriptionId() {
+    return Math.random().toString(36).replace(/[^a-z]+/g, '').substr(0, 12);
+}
